requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
azure-core==1.22.1
azure-identity==1.7.1
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
databases==0.5.5
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.73.0
fastapi-jwt-auth==0.5.0
fastapi-users==9.2.5
fastapi-users-db-sqlalchemy==1.1.0
google-api-core==2.5.0
google-auth==2.6.0
googleapis-common-protos==1.55.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.21.3
httpx-oauth==0.4.1
idna==3.3
install==1.3.5
makefun==1.13.1
msal==1.17.0
msal-extensions==0.3.1
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.1
passlib==1.7.4
portalocker==2.4.0
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydentic==0.0.1.dev3
PyJWT==1.7.1
python-multipart==0.0.5
python-stdnum==1.17
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
